+++ payment/src/test/java/com/example/payment/infrastructure/messaging/kafka/KafkaSellerSettlementPayoutResultEventPublisherTest.java
@@ -5,6 +5,7 @@ import static org.mockito.Mockito.verify;
 import com.example.payment.infrastructure.messaging.kafka.contract.SellerSettlementPayoutResultMessage;
 import com.example.payment.infrastructure.messaging.kafka.contract.SellerSettlementPayoutResultStatus;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 import org.junit.jupiter.api.DisplayName;
@@ -38,7 +39,7 @@ class KafkaSellerSettlementPayoutResultEventPublisherTest {
                 UUID.randomUUID(),
                 settlementId,
                 UUID.randomUUID(),
-                9_000L,
+                BigDecimal.valueOf(9_000L),
                 SellerSettlementPayoutResultStatus.SUCCESS,
                 null,
                 LocalDateTime.of(2026, 4, 1, 3, 10)
